+++ tests/PhpWord/TemplateProcessorTest.php
@@ -187,7 +187,7 @@ final class TemplateProcessorTest extends \PHPUnit\Framework\TestCase
     {
         $templateProcessor = new TemplateProcessor(__DIR__ . '/_files/templates/header-footer.docx');
 
-        $this->assertEquals(array('documentContent', 'headerValue', 'footerValue'), $templateProcessor->getVariables());
+        $this->assertEquals(array('documentContent', 'headerValue:100:100', 'footerValue'), $templateProcessor->getVariables());
 
         $macroNames = array('headerValue', 'documentContent', 'footerValue');
         $macroValues = array('Header Value', 'Document text.', 'Footer Value');
@@ -198,6 +198,83 @@ final class TemplateProcessorTest extends \PHPUnit\Framework\TestCase
         $docFound = file_exists($docName);
         unlink($docName);
         $this->assertTrue($docFound);
+    }
+
+    /**
+     * @covers ::setImageValue
+     * @test
+     */
+    public function testSetImageValue()
+    {
+        $templateProcessor = new TemplateProcessor(__DIR__ . '/_files/templates/header-footer.docx');
+        $imagePath = __DIR__ . '/_files/images/earth.jpg';
+
+        $variablesReplace = array(
+                                'headerValue'       => $imagePath,
+                                'documentContent'   => array('path' => $imagePath, 'width' => 500, 'height' => 500),
+                                'footerValue'       => array('path' => $imagePath, 'width' => 100, 'height' => 50, 'ratio' => false),
+        );
+        $templateProcessor->setImageValue(array_keys($variablesReplace), $variablesReplace);
+
+        $docName = 'header-footer-images-test-result.docx';
+        $templateProcessor->saveAs($docName);
+
+        $this->assertFileExists($docName, "Generated file '{$docName}' not found!");
+
+        $expectedDocumentZip = new \ZipArchive();
+        $expectedDocumentZip->open($docName);
+        $expectedContentTypesXml = $expectedDocumentZip->getFromName('[Content_Types].xml');
+        $expectedDocumentRelationsXml = $expectedDocumentZip->getFromName('word/_rels/document.xml.rels');
+        $expectedHeaderRelationsXml = $expectedDocumentZip->getFromName('word/_rels/header1.xml.rels');
+        $expectedFooterRelationsXml = $expectedDocumentZip->getFromName('word/_rels/footer1.xml.rels');
+        $expectedMainPartXml = $expectedDocumentZip->getFromName('word/document.xml');
+        $expectedHeaderPartXml = $expectedDocumentZip->getFromName('word/header1.xml');
+        $expectedFooterPartXml = $expectedDocumentZip->getFromName('word/footer1.xml');
+        $expectedImage = $expectedDocumentZip->getFromName('word/media/image_rId11_document.jpeg');
+        if (false === $expectedDocumentZip->close()) {
+            throw new \Exception("Could not close zip file \"{$docName}\".");
+        }
+
+        $this->assertNotEmpty($expectedImage, 'Embed image doesn\'t found.');
+        $this->assertContains('/word/media/image_rId11_document.jpeg', $expectedContentTypesXml, '[Content_Types].xml missed "/word/media/image5_document.jpeg"');
+        $this->assertContains('/word/_rels/header1.xml.rels', $expectedContentTypesXml, '[Content_Types].xml missed "/word/_rels/header1.xml.rels"');
+        $this->assertContains('/word/_rels/footer1.xml.rels', $expectedContentTypesXml, '[Content_Types].xml missed "/word/_rels/footer1.xml.rels"');
+        $this->assertNotContains('${documentContent}', $expectedMainPartXml, 'word/document.xml has no image.');
+        $this->assertNotContains('${headerValue}', $expectedHeaderPartXml, 'word/header1.xml has no image.');
+        $this->assertNotContains('${footerValue}', $expectedFooterPartXml, 'word/footer1.xml has no image.');
+        $this->assertContains('media/image_rId11_document.jpeg', $expectedDocumentRelationsXml, 'word/_rels/document.xml.rels missed "media/image5_document.jpeg"');
+        $this->assertContains('media/image_rId11_document.jpeg', $expectedHeaderRelationsXml, 'word/_rels/header1.xml.rels missed "media/image5_document.jpeg"');
+        $this->assertContains('media/image_rId11_document.jpeg', $expectedFooterRelationsXml, 'word/_rels/footer1.xml.rels missed "media/image5_document.jpeg"');
+
+        unlink($docName);
+
+        // dynamic generated doc
+        $testFileName = 'images-test-sample.docx';
+        $phpWord = new \PhpOffice\PhpWord\PhpWord();
+        $section = $phpWord->addSection();
+        $section->addText('${Test:width=100:ratio=true}');
+        $objWriter = \PhpOffice\PhpWord\IOFactory::createWriter($phpWord, 'Word2007');
+        $objWriter->save($testFileName);
+        $this->assertFileExists($testFileName, "Generated file '{$testFileName}' not found!");
+
+        $resultFileName = 'images-test-result.docx';
+        $templateProcessor = new \PhpOffice\PhpWord\TemplateProcessor($testFileName);
+        unlink($testFileName);
+        $templateProcessor->setImageValue('Test', $imagePath);
+        $templateProcessor->setImageValue('Test1', $imagePath);
+        $templateProcessor->setImageValue('Test2', $imagePath);
+        $templateProcessor->saveAs($resultFileName);
+        $this->assertFileExists($resultFileName, "Generated file '{$resultFileName}' not found!");
+
+        $expectedDocumentZip = new \ZipArchive();
+        $expectedDocumentZip->open($resultFileName);
+        $expectedMainPartXml = $expectedDocumentZip->getFromName('word/document.xml');
+        if (false === $expectedDocumentZip->close()) {
+            throw new \Exception("Could not close zip file \"{$resultFileName}\".");
+        }
+        unlink($resultFileName);
+
+        $this->assertNotContains('${Test}', $expectedMainPartXml, 'word/document.xml has no image.');
     }
 
     /**